opt.: handle esc btn in ssh page Fixes #736
