+++ lib/view/page/ssh/page/keyboard.dart
@@ -1,0 +1,20 @@
+part of 'page.dart';
+
+extension _Keyboard on SSHPageState {
+  void _handleEscKeyOrBackButton() {
+    _terminal.keyInput(TerminalKey.escape);
+    HapticFeedback.lightImpact();
+  }
+
+  // Handle hardware keyboard events, particularly ESC key
+  bool _handleKeyEvent(KeyEvent event) {
+    if (event is KeyDownEvent) {
+      if (event.logicalKey == LogicalKeyboardKey.escape) {
+        // Prevent default behavior and send to terminal
+        _handleEscKeyOrBackButton();
+        return true; // Mark as handled so it doesn't propagate
+      }
+    }
+    return false; // Let other handlers process this event
+  }
+}
