+++ lib/view/page/systemd.dart
@@ -4,7 +4,7 @@ import 'package:server_box/core/route.dart';
 import 'package:server_box/data/model/server/server_private_info.dart';
 import 'package:server_box/data/model/server/systemd.dart';
 import 'package:server_box/data/provider/systemd.dart';
-import 'package:server_box/view/page/ssh/page.dart';
+import 'package:server_box/view/page/ssh/page/page.dart';
 
 final class SystemdPage extends StatefulWidget {
   final SpiRequiredArgs args;
